# swagger.yaml
openapi: 3.0.0
info:
  title: Hotel Management System API
  version: 1.0.0
  description: >
    একটি সম্পূর্ণ RESTful API হোটেল ম্যানেজমেন্ট সিস্টেমের জন্য,
    যেখানে রুম বুকিং, ইউজার ম্যানেজমেন্ট, পেমেন্ট, স্টাফ টাস্ক, রিপোর্ট ও নোটিফিকেশন অন্তর্ভুক্ত।
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, receptionist, customer, staff]
        status:
          type: string
          enum: [active, banned]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, receptionist, customer, staff]
        status:
          type: string
          enum: [active, banned]
    
    # Rooms Schemas
...
security:
  - BearerAuth: []
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: সকল ইউজারের তালিকা দেখুন (শুধুমাত্র অ্যাডমিন)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: ইউজার তালিকা সফলভাবে রিটার্ন হয়েছে
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      tags:
        - Users
      summary: নতুন ইউজার তৈরি (শুধুমাত্র অ্যাডমিন)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: ইউজার সফলভাবে তৈরি হয়েছে
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...